extract and test secret key resolver
