+++ app/models/terminal_executor.rb
@@ -43,7 +43,7 @@ class TerminalExecutor
   def resolve_secrets(command)
     deploy_groups = @deploy.try(:stage).try(:deploy_groups) || []
     project = @deploy.try(:project)
-    resolver = Kubernetes::ResourceTemplate::SecretKeyResolver.new(project, deploy_groups)
+    resolver = Samson::Secrets::KeyResolver.new(project, deploy_groups)
 
     result = command.gsub(/\b#{SECRET_PREFIX}(#{SecretStorage::SECRET_KEY_REGEX})\b/) do
       key = $1
