+++ plugins/kubernetes/test/models/kubernetes/resource_template_test.rb
@@ -162,35 +162,6 @@ describe Kubernetes::ResourceTemplate do
         e.message.must_include "bar (tried: production/foo/pod1/bar"
         e.message.must_include "baz (tried: production/foo/pod1/baz"
       end
-
-      describe "scopes" do
-        before do
-          SecretStorage.stubs(:read).returns(true)
-        end
-
-        it "looks up by only the key" do
-          SecretStorage.expects(:read_multi).returns('global/global/global/bar' => 'xyz')
-          template.to_hash[:spec][:template][:metadata][:annotations][:'secret/FOO'].
-            must_equal "global/global/global/bar"
-        end
-
-        it "fails when unable to find by onl key" do
-          SecretStorage.expects(:read_multi).returns({})
-          e = assert_raises(Samson::Hooks::UserError) { template.to_hash }
-          # order is important here and should not change
-          priority = [
-            "production/foo/pod1/bar",
-            "global/foo/pod1/bar",
-            "production/global/pod1/bar",
-            "global/global/pod1/bar",
-            "production/foo/global/bar",
-            "global/foo/global/bar",
-            "production/global/global/bar",
-            "global/global/global/bar"
-          ]
-          e.message.must_equal "Failed to resolve secret keys:\n\tbar (tried: #{priority.join(", ")})"
-        end
-      end
     end
 
     describe "daemon_set" do
